% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_last_row_to_MNE.R
\name{add_last_row_to_MNE}
\alias{add_last_row_to_MNE}
\title{Add Last Row to MNE}
\usage{
add_last_row_to_MNE(MNE, new_site_data)
}
\arguments{
\item{hpa}{The hpa dataset.}

\item{NewSite}{The new site data.}
}
\value{
A combined MNE dataset.
}
\description{
Combines two datasets into one MNE dataset.
}
