% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMNEtoVariables.R
\name{convertMNEtoVariables}
\alias{convertMNEtoVariables}
\title{Convert MNE to Variables}
\usage{
convertMNEtoVariables(MNE, individuals = NULL)
}
\arguments{
\item{MNE}{The MNE dataset.}
}
\value{
A list containing data frames of MNE, MAU, PMAU, and AcReMAU.
}
\description{
Converts MNE data into MAU, PMAU, and AcReMAU variables.
}
