% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hpa}
\alias{hpa}
\title{Hominin Part Abundance Dataset}
\format{
A data frame with X rows and Y variables:
\describe{
  \item{column1}{Description of column1}
  \item{column2}{Description of column2}
  ...
}
}
\source{
Generated by \code{generate_newdata(5)}
}
\usage{
hpa
}
\description{
A dataset containing hominin part abundance data.
}
\keyword{datasets}
