% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tapho_data.R
\name{load.tapho.data}
\alias{load.tapho.data}
\title{Load Taphonomy Data from Excel}
\usage{
load.tapho.data(file_path)
}
\arguments{
\item{file_path}{The path to the Excel file.}
}
\value{
A tibble containing the taphonomy data.
}
\description{
This function loads taphonomy data from an Excel file and returns it as a tibble.
}
\examples{
\dontrun{
NewSite <- load.tapho.data("path/to/yourfile.xlsx")
}
}
