% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abundance_function.R
\name{plot.abundance.function}
\alias{plot.abundance.function}
\title{Plot Abundance Function}
\usage{
\method{plot}{abundance.function}(NewSite, hpa)
}
\arguments{
\item{NewSite}{The new site data.}

\item{hpa}{The hpa dataset.}
}
\value{
A plotly figure.
}
\description{
A wrapper function to process data and generate plots.
}
